+++ interfaces/HO-Portal/src/environments/environment.ts
@@ -5,6 +5,8 @@
 export const environment = {
   production: false,
 
+  defaultLocale: 'en',
+
   // APIs
   url_121_service_api: 'http://localhost:3000/api',