+++ interfaces/PA-App/src/app/user-menu/user-menu.component.spec.ts
@@ -1,6 +1,10 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { LoadingController, PopoverController } from '@ionic/angular';
+import {
+  AlertController,
+  LoadingController,
+  ModalController,
+} from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
 import { MockPaDataService } from 'src/app/mocks/padata.service.mock';
 import { PaDataService } from 'src/app/services/padata.service';
@@ -17,11 +21,14 @@ describe('UserMenuComponent', () => {
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       providers: [
         {
+          provide: ModalController,
+        },
+        {
           provide: PaDataService,
           useValue: MockPaDataService,
         },
         {
-          provide: PopoverController,
+          provide: AlertController,
         },
         {
           provide: LoadingController,