+++ interfaces/Referral-App/src/app/mocks/api.offers.mock.ts
@@ -1,234 +1,198 @@
 export default [
   {
     offerID: 1,
-    offerName: { en: 'Wereld Huis' },
+    offerName: 'Wereld Huis',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.wereldhuis.org',
     subCategoryID: 1,
     categoryID: 1,
   },
   {
     offerID: 2,
-    offerName: { en: 'Boost' },
+    offerName: 'Boost',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.boostamsterdam.nl',
     subCategoryID: 1,
     categoryID: 1,
   },
   {
     offerID: 3,
-    offerName: { en: 'Hvoquerido' },
+    offerName: 'Hvoquerido',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.hvoquerido.nl',
     subCategoryID: 2,
     categoryID: 1,
   },
   {
     offerID: 4,
-    offerName: { en: 'Leger des Heils' },
+    offerName: 'Leger des Heils',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.legerdesheils.nl',
     subCategoryID: 2,
     categoryID: 1,
   },
   {
     offerID: 5,
-    offerName: { en: 'Gemeente Amsterdam' },
+    offerName: 'Gemeente Amsterdam',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.amsterdam.nl',
     subCategoryID: 2,
     categoryID: 1,
   },
   {
     offerID: 6,
-    offerName: { en: 'Ask V' },
+    offerName: 'Ask V',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.amsterdam.nl',
     subCategoryID: 2,
     categoryID: 1,
   },
   {
     offerID: 7,
-    offerName: { en: 'Harriet Tubman Huis' },
+    offerName: 'Harriet Tubman Huis',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.harriettubmanhuis.nl',
     subCategoryID: 3,
     categoryID: 1,
   },
   {
     offerID: 8,
-    offerName: { en: 'Janette Noelhuis' },
+    offerName: 'Janette Noelhuis',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.noelhuis.nl',
     subCategoryID: 3,
     categoryID: 1,
   },
   {
     offerID: 9,
-    offerName: { en: 'Vluchtlingen Werk' },
+    offerName: 'Vluchtlingen Werk',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.vluchtlingenwerk.nl',
     subCategoryID: 4,
     categoryID: 2,
   },
   {
     offerID: 10,
-    offerName: { en: 'Juridischolket' },
+    offerName: 'Juridischolket',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.juridischolket.nl',
     subCategoryID: 4,
     categoryID: 2,
   },
   {
     offerID: 11,
-    offerName: { en: 'Ask V' },
+    offerName: 'Ask V',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.askv.nl',
     subCategoryID: 4,
     categoryID: 2,
   },
   {
     offerID: 12,
-    offerName: { en: 'Gemeente Amsterdam' },
+    offerName: 'Gemeente Amsterdam',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.amsterdam.nl',
     subCategoryID: 4,
     categoryID: 2,
   },
   {
     offerID: 13,
-    offerName: { en: 'Advocaten Spreekuur' },
+    offerName: 'Advocaten Spreekuur',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.advocatenspreekuur.nl',
     subCategoryID: 4,
     categoryID: 2,
   },
   {
     offerID: 14,
-    offerName: { en: 'Dokters van de Wereld' },
+    offerName: 'Dokters van de Wereld',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.doktersvandewereld.org',
     subCategoryID: 5,
     categoryID: 3,
   },
   {
     offerID: 15,
-    offerName: { en: 'Oude Zijds 100' },
+    offerName: 'Oude Zijds 100',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.oudezijds100.nl',
     subCategoryID: 5,
     categoryID: 3,
   },
   {
     offerID: 16,
-    offerName: { en: 'Spoedeisende Psychiatrie' },
+    offerName: 'Spoedeisende Psychiatrie',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.spoedeisendepsychiatrieamsterdam.nl',
     subCategoryID: 6,
     categoryID: 3,
   },
   {
     offerID: 17,
-    offerName: { en: 'Oude Zijds 100' },
+    offerName: 'Oude Zijds 100',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.oudezijds100.nl',
     subCategoryID: 6,
     categoryID: 3,
   },
   {
     offerID: 18,
-    offerName: { en: 'i-psy' },
+    offerName: 'i-psy',
     offerIcon:
       'https://www.510.global/wp-content/uploads/2017/07/510-LOGO-WEBSITE-01.png',
-    offerDescription: {
-      en:
-        'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
-    },
+    offerDescription:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
     offerLink: 'www.i-psy.nl',
     subCategoryID: 6,
     categoryID: 3,