+++ interfaces/Referral-App/src/app/services/spreadsheet.service.ts
@@ -1,10 +1,14 @@
 import { Injectable } from '@angular/core';
+import helpMock from 'src/app/mocks/help.mock';
+import referralPageDataMock from 'src/app/mocks/referral-page-data.mock';
 import { Category } from 'src/app/models/category.model';
 import { Help } from 'src/app/models/help.model';
 import { Offer } from 'src/app/models/offer.model';
+import { ReferralPageData } from 'src/app/models/referral-page-data';
+import { SeverityLevel } from 'src/app/models/severity-level.model';
 import { SubCategory } from 'src/app/models/sub-category.model';
+import { LoggingService } from 'src/app/services/logging.service';
 import { environment } from 'src/environments/environment';
-import helpMock from '../mocks/help.mock';
 
 @Injectable({
   providedIn: 'root',
@@ -15,36 +19,36 @@ export class SpreadsheetService {
   private categorySheetIndex = 2;
   private subCategorySheetIndex = 3;
   private offerSheetIndex = 1;
-  private helpSheetIndex = 5;
+  private helpPageSheetIndex = 5;
+  private referralPageSheetIndex = 6;
 
-  constructor() {}
+  constructor(public loggingService: LoggingService) {}
 
   static readCellValue(row, key): string {
     return row[key].$t;
   }
 
-  convertCategoryRowToCategoryObject(categoryRow): Category {
+  convertCategoryRowToCategoryObject = (categoryRow): Category => {
     return {
       categoryID: Number(
         SpreadsheetService.readCellValue(categoryRow, 'gsx$categoryid'),
       ),
-      categoryName: {
-        en: SpreadsheetService.readCellValue(categoryRow, 'gsx$categoryname'),
-      },
+      categoryName: SpreadsheetService.readCellValue(
+        categoryRow,
+        'gsx$categoryname',
+      ),
       categoryIcon: SpreadsheetService.readCellValue(
         categoryRow,
         'gsx$categoryicon',
       ),
-      categoryDescription: {
-        en: SpreadsheetService.readCellValue(
-          categoryRow,
-          'gsx$categorydescription',
-        ),
-      },
+      categoryDescription: SpreadsheetService.readCellValue(
+        categoryRow,
+        'gsx$categorydescription',
+      ),
     };
-  }
+  };
 
-  getCategories(): Promise<Category[]> {
+  getCategories = (): Promise<Category[]> => {
     return fetch(
       `${this.spreadsheetURL}/${this.spreadsheetId}/${this.categorySheetIndex}` +
         '/public/values?alt=json',
@@ -53,39 +57,38 @@ export class SpreadsheetService {
       .then((response) => {
         return response.feed.entry.map(this.convertCategoryRowToCategoryObject);
       })
-      .catch((_) => {
+      .catch((error) => {
+        if (this.loggingService) {
+          this.loggingService.logException(error, SeverityLevel.Critical);
+        }
         return [];
       });
-  }
+  };
 
-  convertSubCategoryRowToSubCategoryObject(subCategoryRow): SubCategory {
+  convertSubCategoryRowToSubCategoryObject = (subCategoryRow): SubCategory => {
     return {
       subCategoryID: Number(
         SpreadsheetService.readCellValue(subCategoryRow, 'gsx$subcategoryid'),
       ),
-      subCategoryName: {
-        en: SpreadsheetService.readCellValue(
-          subCategoryRow,
-          'gsx$subcategoryname',
-        ),
-      },
+      subCategoryName: SpreadsheetService.readCellValue(
+        subCategoryRow,
+        'gsx$subcategoryname',
+      ),
       subCategoryIcon: SpreadsheetService.readCellValue(
         subCategoryRow,
         'gsx$subcategoryicon',
       ),
-      subCategoryDescription: {
-        en: SpreadsheetService.readCellValue(
-          subCategoryRow,
-          'gsx$subcategorydescription',
-        ),
-      },
+      subCategoryDescription: SpreadsheetService.readCellValue(
+        subCategoryRow,
+        'gsx$subcategorydescription',
+      ),
       categoryID: Number(
         SpreadsheetService.readCellValue(subCategoryRow, 'gsx$categoryid'),
       ),
     };
-  }
+  };
 
-  getSubCategories(): Promise<SubCategory[]> {
+  getSubCategories = (): Promise<SubCategory[]> => {
     return fetch(
       `${this.spreadsheetURL}/${this.spreadsheetId}/${this.subCategorySheetIndex}` +
         '/public/values?alt=json',
@@ -96,23 +99,25 @@ export class SpreadsheetService {
           this.convertSubCategoryRowToSubCategoryObject,
         );
       })
-      .catch((_) => {
+      .catch((error) => {
+        if (this.loggingService) {
+          this.loggingService.logException(error, SeverityLevel.Critical);
+        }
         return [];
       });
-  }
+  };
 
-  convertOfferRowToOfferObject(offerRow): Offer {
+  convertOfferRowToOfferObject = (offerRow): Offer => {
     return {
       offerID: Number(
         SpreadsheetService.readCellValue(offerRow, 'gsx$offerid'),
       ),
-      offerName: {
-        en: SpreadsheetService.readCellValue(offerRow, 'gsx$name'),
-      },
+      offerName: SpreadsheetService.readCellValue(offerRow, 'gsx$name'),
       offerIcon: SpreadsheetService.readCellValue(offerRow, 'gsx$icon'),
-      offerDescription: {
-        en: SpreadsheetService.readCellValue(offerRow, 'gsx$whatservice'),
-      },
+      offerDescription: SpreadsheetService.readCellValue(
+        offerRow,
+        'gsx$whatservice',
+      ),
       offerLink: SpreadsheetService.readCellValue(
         offerRow,
         'gsx$linktowebsite',
@@ -149,9 +154,9 @@ export class SpreadsheetService {
         SpreadsheetService.readCellValue(offerRow, 'gsx$categoryid'),
       ),
     };
-  }
+  };
 
-  getOffers(): Promise<Offer[]> {
+  getOffers = (): Promise<Offer[]> => {
     return fetch(
       `${this.spreadsheetURL}/${this.spreadsheetId}/${this.offerSheetIndex}` +
         '/public/values?alt=json',
@@ -162,38 +167,86 @@ export class SpreadsheetService {
           .map(this.convertOfferRowToOfferObject)
           .filter((offer) => offer.offerVisible);
       })
-      .catch((_) => {
+      .catch((error) => {
+        if (this.loggingService) {
+          this.loggingService.logException(error, SeverityLevel.Critical);
+        }
         return [];
       });
-  }
+  };
 
-  convertHelpRowToHelpObject(helpRows): Help {
+  convertHelpRowToHelpObject = (helpRows): Help => {
     return {
-      helpIcon: SpreadsheetService.readCellValue(helpRows[0], 'gsx$value'),
-      helpText: {
-        en: SpreadsheetService.readCellValue(helpRows[1], 'gsx$value'),
-      },
-      helpPhoneLabel: {
-        en: SpreadsheetService.readCellValue(helpRows[2], 'gsx$value'),
-      },
-      helpPhone: SpreadsheetService.readCellValue(helpRows[3], 'gsx$value'),
-      helpWhatsApp: SpreadsheetService.readCellValue(helpRows[4], 'gsx$value'),
-      helpFacebook: SpreadsheetService.readCellValue(helpRows[5], 'gsx$value'),
-      helpTwitter: SpreadsheetService.readCellValue(helpRows[6], 'gsx$value'),
+      helpPageTitle: SpreadsheetService.readCellValue(helpRows[0], 'gsx$value'),
+      helpIcon: SpreadsheetService.readCellValue(helpRows[1], 'gsx$value'),
+      helpText: SpreadsheetService.readCellValue(helpRows[2], 'gsx$value'),
+      helpPhoneLabel: SpreadsheetService.readCellValue(
+        helpRows[3],
+        'gsx$value',
+      ),
+      helpPhone: SpreadsheetService.readCellValue(helpRows[4], 'gsx$value'),
+      helpWhatsApp: SpreadsheetService.readCellValue(helpRows[5], 'gsx$value'),
+      helpFacebook: SpreadsheetService.readCellValue(helpRows[6], 'gsx$value'),
+      helpTwitter: SpreadsheetService.readCellValue(helpRows[7], 'gsx$value'),
     };
-  }
+  };
 
-  getHelp(): Promise<Help> {
+  getHelp = (): Promise<Help> => {
     return fetch(
-      `${this.spreadsheetURL}/${this.spreadsheetId}/${this.helpSheetIndex}` +
+      `${this.spreadsheetURL}/${this.spreadsheetId}/${this.helpPageSheetIndex}` +
         '/public/values?alt=json',
     )
       .then((response) => response.json())
       .then((response) => {
         return this.convertHelpRowToHelpObject(response.feed.entry);
       })
-      .catch((_) => {
+      .catch((error) => {
+        if (this.loggingService) {
+          this.loggingService.logException(error, SeverityLevel.Critical);
+        }
         return helpMock;
       });
-  }
+  };
+
+  convertReferralPageDataRowToReferralPageDataObject = (
+    referralPageDataRows,
+  ): ReferralPageData => {
+    return {
+      referralPageTitle: SpreadsheetService.readCellValue(
+        referralPageDataRows[0],
+        'gsx$value',
+      ),
+      referralHelpButtonLabel: SpreadsheetService.readCellValue(
+        referralPageDataRows[1],
+        'gsx$value',
+      ),
+      referralPageGreeting: SpreadsheetService.readCellValue(
+        referralPageDataRows[2],
+        'gsx$value',
+      ),
+      referralPageInstructions: SpreadsheetService.readCellValue(
+        referralPageDataRows[3],
+        'gsx$value',
+      ),
+    };
+  };
+
+  getReferralPageData = (): Promise<ReferralPageData> => {
+    return fetch(
+      `${this.spreadsheetURL}/${this.spreadsheetId}/${this.referralPageSheetIndex}` +
+        '/public/values?alt=json',
+    )
+      .then((response) => response.json())
+      .then((response) => {
+        return this.convertReferralPageDataRowToReferralPageDataObject(
+          response.feed.entry,
+        );
+      })
+      .catch((error) => {
+        if (this.loggingService) {
+          this.loggingService.logException(error, SeverityLevel.Critical);
+        }
+        return referralPageDataMock;
+      });
+  };
 }