+++ interfaces/Referral-App/src/app/shared/shared.module.ts
@@ -2,7 +2,6 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
-import { TranslateModule } from '@ngx-translate/core';
 import { CategoryComponent } from 'src/app/components/category/category.component';
 import { OfferComponent } from 'src/app/components/offer/offer.component';
 import { SubCategoryComponent } from 'src/app/components/sub-category/sub-category.component';
@@ -17,12 +16,11 @@ import { SubCategoryFilterPipe } from 'src/app/pipes/sub-category-filter.pipe';
     CategoryFilterPipe,
     SubCategoryFilterPipe,
   ],
-  imports: [CommonModule, IonicModule, FormsModule, TranslateModule.forChild()],
+  imports: [CommonModule, IonicModule, FormsModule],
   exports: [
     CategoryComponent,
     SubCategoryComponent,
     OfferComponent,
-    TranslateModule,
     CategoryFilterPipe,
     SubCategoryFilterPipe,
   ],