+++ services/121-service/src/programs/fsp/dto/export-voucher.dto.ts
@@ -1,0 +1,12 @@
+import { ApiModelProperty } from '@nestjs/swagger';
+
+import { Length, IsNumber } from 'class-validator';
+
+export class ExportVoucherDto {
+  @ApiModelProperty({ example: 'did:sov:2wJPyULfLLnYTEFYzByfUR' })
+  @Length(29, 30)
+  public readonly did: string;
+  @ApiModelProperty({ example: 1 })
+  @IsNumber()
+  public readonly installment: number;
+}